test(chore): Renamed and reorganized files and added files flow diagram
Renamed `startupSteps.py` to `commonInitSteps.py`.
Cleanup `Processes` folder (not used).
Moved `walletInitSteps.py` inside the own suite steps.
Added files flow diagram.

Closes #8353
