+++ test/ui-test/doc/status_ui_test_app_files_diagram.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2022-11-21T12:17:31.812Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.3.0 Chrome/104.0.5112.114 Electron/20.1.3 Safari/537.36" etag="2gZtmRdm4FSJ7847yAbF" version="20.3.0" type="device"><diagram id="BqCeokwvJdPy5Db-PBBX" name="Page-1">5Vxbd9pIs/01fkyWrr48YktxNCsSIRbJyC/fwkKWBdh4QA5Iv/7sXdWAgTjJZOyZJCd5gJZafanetWtXgTlwz26X57PB/U08HRaTA8caLg/c4MBxbNs6xAuvNHrl6ORIL5Szamg6bS5cVG1hLlrm6kM1LOZbHevpdFJX99sX8+ndXZHXW9cGs9l0sd3tejrZnvV+UBZ7Fy7ywWT/6qdqWN/o1WPnaHP9bVGVN6uZ7cMTvXM7WHU2O5nfDIbTxaNLbnjgns2m01rf3S7PigmNt7KLPvfmibvrhc2Ku/p7Hng1zl6djctqNPvfWTQfLmadv/545ZrVfh5MHsyOD5zDCQY8nd8P7vC+5Pu0mNfodTaYF6vbV7PVzf9/V2Dkx9Yx51s3K9Asbqq6uLgf5Gwv4BjodFPfTtCyadkVEti4nt7VBvb2MdvVZHI2nUxnMpR77fM/H6pn03Hx6M6h/DMjPLqu/8z1N4PbakLX+1jMhoO7wWpCs1YDos/FrC6WT+LKXqMVbl5Mb4t61qCLecA9NgA3Hu6s2ouNvxwdmWs3j3zF9czFgfHRcj32BsZ4Y5D8N1DN559A9dXWSR3+9UD3E5u8msspdNDB9u6Xetzm/uqkL/56qOY36HEa4GmruCuru5/WI2C6qychOps+3A2LoQHhNwD7DCjxnW2UuCf+Hkr8ky+gxD55KZTYeyC5Gg7/dzOdjuev75s9i8Fr7/n2bloX3zbZ1SAfl2Lk7kM9IUr0+nAwG3fxVFXTENZrS3xb/d95LlufbNvaO9z3yOMvmdp6MVMf7tl6Xhf3v7idTw5/Ojsf7dm5uqtA97+8rd2Tw5/N1sd7tq4hk15fF4P6YVb8wrY+OvJ/Nlvvy9SLfFYUd782qHcNfej914ZeRelHhg5mFTb5exn62PnPDe3uGXrPvsXdsMMUFq18MpjPq3zbosWyqv8Um/mmlT26E3Dv1qrRmMZG91li+/nNWgRqhrHKch1eGRd1fmP6rrMWaz9LGQ6K4+v8i1lKflxcXT+djeAAZ82fq5WykT1uyCaON81mBQppvi9mFY6jmD3a6e41tWox3Evzd7ADy08fZnnxTRWzD7JZMRnUcJKt0b6EGPPo+2l1Vz+SEScnr3d4YBd2ujTz3OMcf2eoY397oCNrZzX1YFYW9d5AgNmgedTtnh3mTy8ZfmHtTLRVfMAbHXLjIGt7/QOf8fZ8Jrr7DLzN910H5/1ucFVMCPkCbDK4klvWtgcNJlV5R/cCFoiYU7JMlQ8mHXPjthoO+eBT8D14XHwycxysSz7f4KyvEMOzg8x1t87r1W7uPb2+nhf1wS6tPcep+c/JdM7vQHWvLdvdorvX/tEW420TnvUU330/t6nff1NhPTvsPHeX2/zDH+Q2DrWrl452xnomevO8bXY7NEXrl2W3/Rz5N2Q3/0Vg9sp27W1+c4/+NX7bz7n/Nr/9vFT13ars0KjQp3QZR3sF5tsicOv14Vcp/KUF3feS49GLoNbfqZyva6R/lxoPraOvD/QELz6bC+yXQvrz34u3XgYB3jZp2TsDvCBpfelzv2dJP62vOvRPR3S/ZMZ4aB/uSiHvR7ljV+t4u/rspVJG99D6Wwvb6f8yIszdr3/9bkz2zEJ/zV3/WkK5knq/k+B6ds75HlVz/DLk5OyTk/ujwsbZETa7ZYsXFjauswe1oLiu7n4rRnCfqJ3/QozwrMX0Hy4x/dTM8ex+fnTs7ZR2jna983vdnEM99QnNv+Xp+7Vl6MLr6ez2gCvZ/Y7NIK+r6d1vxQIvU3f+N1ngWQvNRz+a1LwsCzwqIh+t6y1rovpKAflno49j79s+/90fe3nfpKJnymK+sOrVVC+bl+wXh8/pRdZwUA9+Jxb6p/Xhf/ShvLef/e3ZdvWlh+pWvjm+tpnY+/10XjEw4P7VtK6nt+gw4Y3T9ZcfVk49LK4HDxJVdq1eT/nNicH8Xr/Pfl0tyS2nMmFnddVaXeFQxIDb0abzZv65PHBOlzhm5+z928S5bE69q0/Lh7y1qsHbD1YeTD+/c4fusPHduPE/57f553jUWcRnJ+3wNq+it8P7y7cfpu8voiZJoyo6v5kMPg2nw8Cq4lG/iapTZ/Dpo9u7PfHQZxEFnVKuB5EdVxHmfv/29GZ4XpaXt5P5FZ6K2g9BL42tXhWVl2inabSMG8+L07KM09h/N8q97qdF/W4UOclZZ/lu1HfjC+shbkM3CsKHblC6WYN2Gtnoa8V838Y+7yWjuHl34dlJEJVx5TlxMOY4D9209HDdSoKc1xfJiNdzJ6k6Le4vuzJ+p+UYcdr30deR9VSeFad5HV/gmTQssU43bpO53G8zHSuYjNBu4jbC/GPOY8tago6L9y7maN6NQot7SC64nx6uWXgNW+wP98ZecsZ27uG+hf06ej+2dX3od2Gxn4X7Dp73k8aypN10YK/S7crzYz6P/rGHNu5HNsZ3eB/j4XXc0J7xmdckQVknF14by3i5zfmwRz8ZRTyDh3iU1XE6fkjSnDbxsCackeXi1ccztpzNhbVMsP84zcouxuqmeGbUgx37Fp7R9VRcR+joviMfz2CsCOvlvuNlzHkr2m5MW3L+Gjb2E86NfbxL84ek7em5BDiHFK/Yv55T6V3JvuOFnE8F+456aHcanieec2D/RRzENcZuBSdcVxphjLDVe2PeazBXLXZ2Yp61laQ9tqVvN8g4zjIJenVy5i26Qb/m2jG32FrONQgdYFLmToilFHbFWrtpRAw03SDkPC5si2d7jay1FUxi3L6ZM1fMNxYx3Cajj3OOkYxk/jYOhjcJ+53R/j1LPA12hadhP7BvGi7ejTI7/rSgr9BejeKwz3Va8r7t4Qx7y6zxHPgR1pnp9SDH9diJBauxmzSe1U1v5nImbQ84CP1MxiyBt2ih5xfxTMEaxFzP4h7iFs8AR7jeJJ/CZdyWwCpslvaIK6+L84TNde1nHbAFx44xd2Qlt1j3aIxnO0s8s0zkfEruA/vtLGh7jId2p8H64Xc9eU/7JuLnYC3hhY6nZ/tmLtzR8D1wTVuP+nymzVpgPO1h3R27Kz7St3gmsIcDeyyIBTnXtA/OifFa8hwtjoO1tfB1ntlSOCAVznAEV2kI3wdmR3KmPvyX+7DjIFN/GwHn4KOkhe2AceBiyf4x8If9WUkbiv+BR1o9zxBn0Vl2z+gbwFCaoW/k8syBdfhHTEx7tGM3jTEm7qc51pCh3bc4B9jYF4wAv7RFQl/j+V+QE0KnK3gKG+yXtndixYZichTSbq3MXXleQk7EvmAHrrfBfjB+CdtnHu7b5FXpLzwdNeoD4+bqDGOmcs+P0zcj4cAL+ksEXwX3t+KjsG+p59RmXCsweHkDLifnl6tXxU2nJK+BP4Q/u5/IWTH9xOd5JMSM+M4Ye8ta8alRr6VvJMIBXF/m0y9i4dq+nTjkrNKjvc3rwvgYsa7jBbGnMWjcbF4Fy0vja2pr4Kaoos/vR8tF9ueHaXR+eX91vjiJKsQ9CzwTMpL2TqIx+C/tA18RnwdeyAf0+4g+5Kqv9xf0c+XYUvEz6ljF2HLW19Ix40bbPV/Qxg1wJDErTj8wNnn0C2DSFowGPWMbjJ0KtlvhuXTcZin5gdF79ap74/3EcGX3nByGmAP/6AbEIO0geHdNHy9mG3zXTcOa58M+wKmv84e28ZVG8Ytz42sbmX6Z4CY26wJfOGb+VvqluTyH83PNODIuMIPnGX8QQ2FT3PfEH1rEYfpB22vM/OI3SZBprIEvw6cb2IhY0ZgkfPvmpqt2ASeMiTP6s8GExDU/0fjlAWdll36ckmfBFRrbV30b4Un2hc7BdZ+6ATYru7Ke0l7bmZxC23EfQQ4slVjTmGMhtgPbbc/gH36p40HrIHqZOAefBDf1S4mpaVSDNyxgVHVLSv7AfILvHuMM+K9syG+Yy2GsxbOucHJbWso/wJ4Dn0Hs1jNhH55p3xWsSpyMPNEBLXmg4wovpoyRoW/4RmKrxGwZh3uObY3hmeInBZOCVxkncQYN1g2bZoJlnGOpmBaMSQyB7QTDGG+51n2CkagVbgwMtoPIvRJ/jBVDwECWxoxxPrUSdQF4cmPTC8VsLHu7nCv3CRfa0CTKTUGo60pvNB6LHurZm3jMGBU5wutt5kMncJ2Ncp1oBMYcvvqGA13VufESMRPagGeKPY8y+GbPWcVK9dncVt6JbIzbMFYqP4FBb+mXeZMQYzIu4q3orhznwL6hq/EdOlM0M/wLui4RrVQS34aTQ2o80d3U4WL7kcRPL2lFH2kcu1DfUZvkpZ7Nx5HRLMKD6K++jTONAvokYxTx1admXML2Yts4FYy2xJPwBnyDugG6kD7qdTl+2gFO3twkLeNrDzilZsmsNU+LPXpL7pnnB0zBJ3vkOXA28DzKqFtWfak7yOWN6ONPogcsYh+WJJ5snmMisZO8LrqAfe3up5B4t7EXjJ8Jf270DHms42vf0ErO46VqL/JVLloNfKIaAWtSX4s81Rxlozo5h/4RHU/93jJ+qU6PXdXe8MmKur/jiT5LI43R6ifwm1xzH81PNK9ifiVzd1Q7VxLjWqPjEB8EF9AAkg8Bl4z1cqYLGeeCeU+msV/5xYadSj2rPxCjYxMbqal67cbOMr7JcyKbcRpcib1kjq5hhf1Vrlfa+pq7iNXKU3rfYDJqTC5gZxJ3St/khP52rIossTW05nfGYcZS5nAl+Iq+0VzJmcQL5l4J4wTzD+qdtm8Ty4yjiME+8imP+jaRM87aK8FLjxqUmsfYjhwIe8v+gUX6UCt8w1yVnOeqH/cc9cGOm5wDP41oR4wNvCCn2GgNcraJq+Ae0ebUf9Soox7PBb4ksbd8pOcQ0yZzyVlTridzkO88JMLn0VK5omdrPtNxTV7hSq6rOtBmDkatr3lmuFQN0BEcd6nFTdxWP8i5T5eaLpE4EBpt2FfNjDNSrlCNDRt6YhfgF/wveRHmIjbpz9DGkh/RRzzmOcK9LfNVxoPSMesl/xNfjcmTwGEh4+1SxhW9QP281myN5EziVyXyWYmz5M3V60oDGk0gvIRcjRGH+lhqBK76ZwZeEP8Ez9M/e9SQnthU/KtcdOWMx4jb7Bc3qr0x70XHNX6je2PMbccm74VdRqyL9CXflhwqhVYCn0vexFgF/aLcETfqV2PL+I1t9KwlZx3IOttEtJlgs43TleY2nE5OQ5zUeG1yyDQ0cTVTOweqG8ARtuA9pf2JyY8jqU9U1OE9w3nMOzMdLx1r7sJ5pZZCLl/lGOwfG3/OTc6QqY+M3misBeeuXs351YwL0BPK3YhdEm/T0JG8W3LqvqX1kR54U3Swq7grPakj4T4xDt6Uvup/PXCWRUxqXje6nKuvUa8wrg1HUk9KM6Oz+7XhSslxREPQbzTnWqrfRH5X4o7MZxt/WFKnCSYkDwH3c/1t7poalKlTWXqWso8QOZrUexbAicTneCR6e6HzhHai+5I8WMYLMkv2GuRSc8C4PENq2CaR2gn16GQueTPz4IBaoiO1DdjS+GVWS94muUTG+gni/xuN921f7UKcijYvHWpw6FVLNPiI+lLmVM0u+pJ27LfST3LVPnnbl35pZml/6ofsQfJN2SfsKbkCz2WM53OOtzQ1POrhJXNV0UC0b8P6ncmbwbg6R48xwjK1G0f1K/V8p5G9BVxzzvxZaoYyLvEh+jZm3tyKzhKtnmnuQM2QxrZqZnKo6DtLzymjfeHrfC4EFpXzodXtrvjIuBWeZG0izVyTZ7IeA74Q3dYID+pZ+136Txq7Wm9inhg3qu96yOWp80UzwWd6Er+T0c2NxqZVjFrlEJ1a6zuhxvmR5opSF2pj4yPItaCBNH+OXdFi6QetS7U5a7W24pYcBc691RxaY0n5iD81x1Yb9Kz4z2n1rvWO8/M31uDs9HbwaTl/fxGNkjA5jdsPZ1F1On5/8UeieW3okk+lPtD+MYePS30NPEIbL8GZrnA41sj9d4PLw6g6ZuX5PLnPb5P55cVmhg/hm7A3/hBghlHBGrZUkcd4Oiozd3wSsRKCaI9IX0VvI7SzBdVJ90xUxN2XFMRGXYiS8DSrYMWLbF4uocyN2qJnZx7VA5lD1OuZKDmNuiOjQEe5VMiAGlibrB2iPTYKVap4ysLKjoiUJsOjSm+MWjftRKrdfaPEcjKDIlIjkHpfMPZZMYLnOEOJMH8wutmstnWlQphJRUMYH4jbtMneJppK1h+bjC0vk1V1ld4BJSlZTRtqtTXdZFpaQR5vnpcInTuq3jqtZnGSBa36kklbqQynZSMsfAfWbTtShS1updLddjViQ8V+nErVIr3knhxmHrFE3cwgl1F13W7lvBpmW5FhL9p73JhKic3sBTZpN22Jjq7xJmSwY9rTVFSYvSNbeJsxUi+Tt/c1o5ZGPkQQRJMuK8rIVDQLyuRZzZxC814qHmbNY1cZJfSTTdsi4ySy5lg/FUlXtmM2k9nKwGN71R5IpoTs8C2iGTy8++c9PE2qjRYrlqy6IhNvMn6ikeatqEoqXKp/iZIr3IpaNbhldlcuiYWuKN++so9UpxgZyuUmO2JGGKrCbKnc+6zeeYIrsPCmTTbJBF+ruTUzL0uN+lBAEj16JjtD9BaGiRyT0QAD8olDo5+kjBmBl6xcdi/WbcmwDL4aRuwVvrpSxejBJ6wWKh82GB5eSQU7ZiZnS1XnljgYa8VHlFbfEbXZ5vrpUcv1xa7srZVPJ9TGUu0tJUKpIiZ+c08jSGlJ5QQKOJEx5FMRl1XiVbW0K9W2TVtVUmSy4D6yKq5Bxld12F5CxXeoYFpkLcskWCnhcMFqS0yfFx+VTyiWBl+NKBY53zHVny1KKRAsIFJwPuIg81V9amWD2afcY6VcuakRrlL/KCWSt2Wp9o7W7VgjzgrrjUadUMdk9R9qLhG+yxRTKf1zVV0Qn1mKaoBS0oxCMiR+EuY9fd7xo/Pu+Tt84oJPXP1EiZ/s0I5/jBgRu2fmjNXu5oypkGPHYMGR6jJ9QNbb0U9ZlM9szfhFTXFvXiLquSe+ivfKN+tPCcjTY1HlmzZ5J1sa3sG5yRo8UyEVxTUcRRK7Ylej7TsnZ1TUT3bx7uBf+zGf1dfIH/85/e5Xy5/vg/v9rwTfDW6LX/w3Tw7tbTP/979a4O1/DeVvf8XqZ/vq9X/yx2nP/23M3e9SHf/oV653fjxqb6AX/iKmt//nlL/ZX2B4L/NHQf/Z35L5+6QgP4T0Be6Frert89l217vpXbHj2+bS95/jlwh9m3Me/+Qi29M1j3v7X+BcP2B28X0H/3Vit60dJ1v/Yts3Auju31F8B7GjufmNUT3yzS+1uuH/AQ==</diagram></mxfile>
