+++ test/ui-test/testSuites/suite_communities/tst_communityMessageFlows/bdd_hooks.py
@@ -4,7 +4,7 @@
 sys.path.append(os.path.join(os.path.dirname(__file__), "../../../testSuites/global_shared/"))
 sys.path.append(os.path.join(os.path.dirname(__file__), "../../../src/"))
 
-from steps.startupSteps import *
+import steps.commonInitSteps as init_steps
 
 # Global properties for the specific feature
 _user = "tester123"
@@ -12,8 +12,8 @@ _password = "TesTEr16843/!@00"
 
 @OnFeatureStart
 def hook(context):
-    context_init(context)  
-    signs_up_process_steps(context, _user, _password)
+    init_steps.context_init(context)  
+    init_steps.signs_up_process_steps(context, _user, _password)
 
 @OnFeatureEnd
 def hook(context):
