+++ test/ui-test/testSuites/suite_messaging/config.xml
@@ -1,8 +1,8 @@
 <testconfig version="1.0">
-<information>
-<summary/>
-<description/>
-</information>
-<testsettings/>
-<passwords/>
+    <information>
+        <summary/>
+        <description/>
+    </information>
+    <testsettings/>
+    <passwords/>
 </testconfig>
